moved GitRepository extensions to GitExtensions
