+++ src/GitVersionCore.Tests/Mocks/MockRepository.cs
@@ -30,9 +30,9 @@ namespace GitVersionCore.Tests.Mocks
 
         public BranchCollection Branches { get; set; }
         public TagCollection Tags { get; set; }
-        public RepositoryInformation Info { get; set; }
-
         public Network Network { get; set; }
+        public string Path { get; }
+        public bool IsHeadDetached { get; }
         public int GetNumberOfUncommittedChanges() => 0;
         public Commit FindMergeBase(Commit commit, Commit otherCommit) => throw new NotImplementedException();
         public string ShortenObjectId(Commit commit) => throw new NotImplementedException();
