+++ src/GitVersionCore/Core/Abstractions/IGitRepository.cs
@@ -5,13 +5,14 @@ namespace GitVersion
 {
     public interface IGitRepository : IDisposable
     {
+        string Path { get; }
+        bool IsHeadDetached { get; }
         IGitRepositoryCommands Commands { get; }
         Branch Head { get; }
         CommitCollection Commits { get; }
         BranchCollection Branches { get; }
         TagCollection Tags { get; }
         ReferenceCollection Refs { get; }
-        RepositoryInformation Info { get; }
         Network Network { get; }
         int GetNumberOfUncommittedChanges();
         Commit FindMergeBase(Commit commit, Commit otherCommit);
