+++ src/GitVersionCore/Core/GitPreparer.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using GitVersion.Extensions;
@@ -209,9 +208,7 @@ namespace GitVersion
 
             try
             {
-                var remote = EnsureOnlyOneRemoteIsDefined(repository, log);
-
-                AddMissingRefSpecs(repository, log, remote);
+                var remote = repository.EnsureOnlyOneRemoteIsDefined(log);
 
                 //If noFetch is enabled, then GitVersion will assume that the git repository is normalized before execution, so that fetching from remotes is not required.
                 if (noFetch)
@@ -242,7 +239,7 @@ namespace GitVersion
 
                 var headSha = repository.Refs.Head.TargetIdentifier;
 
-                if (!repository.Info.IsHeadDetached)
+                if (!repository.IsHeadDetached)
                 {
                     log.Info($"HEAD points at branch '{headSha}'.");
                     return;
@@ -295,7 +292,7 @@ namespace GitVersion
                 else if (localBranchesWhereCommitShaIsHead.Count == 0)
                 {
                     log.Info($"No local branch pointing at the commit '{headSha}'. Fake branch needs to be created.");
-                    CreateFakeBranchPointingAtThePullRequestTip(repository, log, authentication);
+                    repository.CreateFakeBranchPointingAtThePullRequestTip(log, authentication);
                 }
                 else
                 {
@@ -409,104 +406,6 @@ Please run `git {GitExtensions.CreateGitLogArgs(100)}` and submit it along with 
             }
 
             repo.Commands.Checkout(localCanonicalName);
-        }
-
-        private static Remote EnsureOnlyOneRemoteIsDefined(IGitRepository repo, ILog log)
-        {
-            var remotes = repo.Network.Remotes;
-            var howMany = remotes.Count();
-
-            if (howMany == 1)
-            {
-                var remote = remotes.Single();
-                log.Info($"One remote found ({remote.Name} -> '{remote.Url}').");
-                return remote;
-            }
-
-            var message = $"{howMany} remote(s) have been detected. When being run on a build server, the Git repository is expected to bear one (and no more than one) remote.";
-            throw new WarningException(message);
-        }
-
-        private static void AddMissingRefSpecs(IGitRepository repo, ILog log, Remote remote)
-        {
-            if (remote.FetchRefSpecs.Any(r => r.Source == "refs/heads/*"))
-                return;
-
-            var allBranchesFetchRefSpec = $"+refs/heads/*:refs/remotes/{remote.Name}/*";
-
-            log.Info($"Adding refspec: {allBranchesFetchRefSpec}");
-
-            repo.Network.Remotes.Update(remote.Name,
-                r => r.FetchRefSpecs.Add(allBranchesFetchRefSpec));
-        }
-
-        private static void CreateFakeBranchPointingAtThePullRequestTip(IGitRepository repo, ILog log, AuthenticationInfo authentication)
-        {
-            var remote = repo.Network.Remotes.Single();
-
-            log.Info("Fetching remote refs to see if there is a pull request ref");
-            var remoteTips = (string.IsNullOrEmpty(authentication.Username) ?
-                    GetRemoteTipsForAnonymousUser(repo, remote) :
-                    GetRemoteTipsUsingUsernamePasswordCredentials(repo, remote, authentication.Username, authentication.Password))
-                .ToList();
-
-            log.Info($"Remote Refs:{System.Environment.NewLine}" + string.Join(System.Environment.NewLine, remoteTips.Select(r => r.CanonicalName)));
-
-            var headTipSha = repo.Head.Tip.Sha;
-
-            var refs = remoteTips.Where(r => r.TargetIdentifier == headTipSha).ToList();
-
-            if (refs.Count == 0)
-            {
-                var message = $"Couldn't find any remote tips from remote '{remote.Url}' pointing at the commit '{headTipSha}'.";
-                throw new WarningException(message);
-            }
-
-            if (refs.Count > 1)
-            {
-                var names = string.Join(", ", refs.Select(r => r.CanonicalName));
-                var message = $"Found more than one remote tip from remote '{remote.Url}' pointing at the commit '{headTipSha}'. Unable to determine which one to use ({names}).";
-                throw new WarningException(message);
-            }
-
-            var reference = refs[0];
-            var canonicalName = reference.CanonicalName;
-            log.Info($"Found remote tip '{canonicalName}' pointing at the commit '{headTipSha}'.");
-
-            if (canonicalName.StartsWith("refs/tags"))
-            {
-                log.Info($"Checking out tag '{canonicalName}'");
-                repo.Commands.Checkout(reference.Target.Sha);
-                return;
-            }
-
-            if (!canonicalName.StartsWith("refs/pull/") && !canonicalName.StartsWith("refs/pull-requests/"))
-            {
-                var message = $"Remote tip '{canonicalName}' from remote '{remote.Url}' doesn't look like a valid pull request.";
-                throw new WarningException(message);
-            }
-
-            var fakeBranchName = canonicalName.Replace("refs/pull/", "refs/heads/pull/").Replace("refs/pull-requests/", "refs/heads/pull-requests/");
-
-            log.Info($"Creating fake local branch '{fakeBranchName}'.");
-            repo.Refs.Add(fakeBranchName, new ObjectId(headTipSha));
-
-            log.Info($"Checking local branch '{fakeBranchName}' out.");
-            repo.Commands.Checkout(fakeBranchName);
-        }
-
-        private static IEnumerable<DirectReference> GetRemoteTipsUsingUsernamePasswordCredentials(IGitRepository repository, Remote remote, string username, string password)
-        {
-            return repository.Network.ListReferences(remote, (url, fromUrl, types) => new UsernamePasswordCredentials
-            {
-                Username = username,
-                Password = password ?? string.Empty
-            }).Select(r => r.ResolveToDirectReference());
-        }
-
-        private static IEnumerable<DirectReference> GetRemoteTipsForAnonymousUser(IGitRepository repository, Remote remote)
-        {
-            return repository.Network.ListReferences(remote).Select(r => r.ResolveToDirectReference());
         }
     }
 }
